Tidy - move tone detection into a separate file
